third_party/blink/web_tests/platform/generic/external/wpt/html/canvas/element/fill-and-stroke-styles/2d.pattern.svgimage.zeroheight-expected.txt

This is a testharness.js-based test.
FAIL Canvas test: 2d.pattern.svgimage.zeroheight assert_equals: ctx.createPattern(img, 'repeat') === null (got [object CanvasPattern][object], expected [object]) expected null but got object "[object CanvasPattern]"
Harness: the test ran to completion.
